backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
python-socketio==5.11.2
python-engineio==4.9.0
eventlet==0.33.3
pymysql==1.1.0
cryptography==41.0.7
psutil==5.9.6
APScheduler==3.10.4
numpy==1.24.0
pandas==2.2.1
scikit-learn==1.4.2
joblib==1.4.0
tensorflow==2.14.0
matplotlib==3.8.3
pickle-mixin==1.0.2
keras==2.14.0
h5py==3.9.0